+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+# Version 0.8.0 (UNRELEASED)
+* Jupyter kernel now shows errors and warnings inline as you type by running
+  cargo check on the backend. Running `evcxr\_jupyter --install` before starting
+  jupyter notebook is recommended. It will auto-update when the evcxr jupyter
+  kernel starts, however that update may not effect the current session.
+* Fixed `:clear` command.
+
 # Version 0.7.0
 * Fixed a number of bugs in tab completion.
 * Documentation improvements. Thanks Ibraheem!